+++ src/components/SelectMonth.vue
@@ -7,6 +7,7 @@
     >
       <template #placeholder>
         <Button type="secondary">
+          <RemixIcon :icon="'calendar-2'" />
           {{ recordsQueryDate[0] }} - {{ recordsQueryDate[1] }}
         </Button>
       </template>
@@ -22,9 +23,10 @@ import dayjs from 'dayjs'
 import { useStore } from '@/store'
 import { getThisMonth } from '@/utils/date'
 import Button from './ui/Button.vue'
+import RemixIcon from './RemixIcon.vue'
 
 export default defineComponent({
-  components: { InputDate, Button },
+  components: { InputDate, Button, RemixIcon },
 
   setup(props) {
     const store = useStore()
@@ -53,6 +55,12 @@ export default defineComponent({
 
 <style lang="less" scoped>
 .select-month {
+  position: fixed;
+  left: 0;
+  right: 0;
+  bottom: calc(var(--tabbar-height) + env(safe-area-inset-bottom));
+  height: var(--home-select-month-height);
+
   & button {
     color: var(--ahnn-foreground);
     width: 100%;
@@ -60,6 +68,13 @@ export default defineComponent({
     backdrop-filter: saturate(180%) blur(5px);
     box-shadow: var(--tabbar-border-top);
     border-radius: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    & svg {
+      margin-right: 0.5rem;
+    }
   }
 }
 </style>