+++ src/hooks/readRecordsFromDB.ts
@@ -1,9 +1,0 @@
-import { useStore } from '@/store'
-import { hooks } from '.'
-
-export function readRecordsFromDB() {
-  hooks.add('appBeforeMount', () => {
-    const store = useStore()
-    store.dispatch('readRecordsByQueryDate')
-  })
-}