+++ src/mixin-events/refreshRecords.ts
@@ -1,0 +1,17 @@
+import { useStore } from '@/store'
+import { computed, watch } from 'vue'
+
+export function eventRefreshRecords() {
+  const store = useStore()
+
+  const recordsQueryDate = computed(() => store.getters.recordsQueryDate)
+  watch(
+    recordsQueryDate,
+    () => {
+      store.dispatch('readRecordsByQueryDate')
+    },
+    {
+      immediate: true,
+    },
+  )
+}