+++ src/utils/date.ts
@@ -26,3 +26,13 @@ export function getCNDayText(val: ConfigType) {
 
   return val
 }
+
+/**
+ * Constants for months in year
+ * ```
+ * [1, 2, 3, ..., 12]
+ * ```
+ */
+export const monthsInYear: number[] = Array(12)
+  .fill('')
+  .map((_, i) => i + 1)